pyproject.toml
[project]
name = "langchain-version-1-x-xcourse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.7",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.7",
    "load-dotenv>=0.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.22",
]
